backend/pyproject.toml
[project]
name = "ai-rag-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "python-dotenv==1.0.1",
  "sqlmodel==0.0.21",
  "qdrant-client==1.9.2",
  "sentence-transformers==3.0.1",
  "httpx==0.27.0",
  "pypdf==4.3.1",
  "tqdm==4.66.5",
  "python-multipart==0.0.9",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["storage*", "tests*", "test*", "docs*", "examples*"]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true
